backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
pydantic[email]==2.9.2

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
psycopg2-binary==2.9.9
alembic==1.13.3

# Queue
celery[redis]==5.4.0
redis==5.1.1

# Storage
boto3==1.35.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# PDF processing
PyMuPDF==1.24.10
ocrmypdf==16.5.0
pytesseract==0.3.13
Pillow==10.4.0
reportlab==4.2.2
pdf2docx==0.5.8
pikepdf==9.7.0

# File validation
python-magic==0.4.27
clamd==1.0.2

# HTTP
httpx==0.27.2
python-multipart==0.0.12
